requirements.txt
alembic==1.15.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
pyclipper==1.3.0.post6
pycparser==2.22
PySocks==1.7.1
python-bidi==0.6.6
python-dotenv==1.1.0
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.3
selenium==4.33.0
shapely==2.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.5.26
torch==2.7.0
torchvision==0.22.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
